fix(town-hall): satisfy python lint for community PR
Move source registration to registry.py so __init__.py stays empty,
rename ambiguous lambda names, and apply autofix formatting.
